requirements.txt
slackclient
spotipy
flask
requests
